requirements.txt
flask==3.0.2
flask-limiter==3.5.0
flask-caching==2.1.0
python-dotenv>=0.19.0
python-osc>=1.7.7
websockets==12.0
gunicorn==21.2.0
requests==2.31.0
PySide6==6.6.1
simple-websocket-server>=0.4.0
Flask>=2.0.1
Flask-SocketIO>=5.1.1
python-socketio>=5.4.0
eventlet>=0.33.0 